requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
lxml==5.3.2
packaging==24.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-vlc==3.0.21203
requests==2.32.3
soupsieve==2.6
typing_extensions==4.13.2
urllib3==2.4.0
